+++ demos/bookstore/app/utils/session.ts
@@ -1,7 +1,7 @@
 import * as path from 'node:path'
 import { fileURLToPath } from 'node:url'
 import { createCookie } from '@remix-run/cookie'
-import { createFileStorage } from '@remix-run/session/file-storage'
+import { createFileSessionStorage } from '@remix-run/session/file-storage'
 
 const __dirname = path.dirname(fileURLToPath(import.meta.url))
 
@@ -21,6 +21,6 @@ export let sessionCookie = createCookie('session', {
  * File-based session storage.
  * Sessions are stored in the app's tmp/sessions directory.
  */
-export let sessionStorage = createFileStorage(
+export let sessionStorage = createFileSessionStorage(
   path.resolve(__dirname, '..', '..', 'tmp', 'sessions'),
 )